+++ service_info/settings/production.py
@@ -10,6 +10,10 @@ DEFAULT_FROM_EMAIL = 'noreply@rescue.org'
 
 SITE_ID = PRODUCTION_SITE_ID
 
+CMS_LANGUAGES[SITE_ID] = CMS_LANGUAGES_FOR_SITE
+PARLER_LANGUAGES[SITE_ID] = PARLER_LANGUAGES_FOR_SITE
+DISQUS_SHORTNAME = 'serviceinfo'
+
 # Uncomment if using celery worker configuration
 if 'BROKER_PASSWORD' in os.environ:
     BROKER_URL = 'amqp://service_info_production:%(BROKER_PASSWORD)s@%(BROKER_HOST)s' \