+++ service_info/settings/staging.py
@@ -38,6 +38,10 @@ ALLOWED_HOSTS = os.environ['ALLOWED_HOSTS'].split(';')
 
 SITE_ID = STAGING_SITE_ID
 
+CMS_LANGUAGES[SITE_ID] = CMS_LANGUAGES_FOR_SITE
+PARLER_LANGUAGES[SITE_ID] = PARLER_LANGUAGES_FOR_SITE
+DISQUS_SHORTNAME = 'serviceinfo-staging'
+
 # Uncomment if using celery worker configuration
 if 'BROKER_PASSWORD' in os.environ:
     CELERY_SEND_TASK_ERROR_EMAILS = True