+++ service_info/static/js/src/component/menu.js
@@ -1,0 +1,13 @@
+var $ = require('jquery');
+window.jQuery = $;
+window.$ = $;
+
+var menu = require('../config').components.menu;
+
+function init () {
+    $(menu.toggle).click(function() {
+        $(menu.container).toggleClass([menu.closed_class, menu.open_class].join(' '));
+    });
+}
+
+module.exports = init;