+++ service_info/static/js/src/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    components: {
+        menu: {
+            toggle: '#menu-toggle'
+            , container: '#menu-container'
+            , closed_class: 'menu-closed'
+            , open_class: 'menu-open'
+        }
+    }
+};