+++ service_info/static/js/src/index.js
@@ -1,0 +1,13 @@
+/*
+  UI COMPONENTS
+*/
+
+/*
+  Mobile menu show/hide
+*/
+require('./component/menu')();
+
+/*
+  Initializing Google Analytics
+*/
+require('../../../../frontend/google-analytics.js')();